requirements.txt
azure-common==1.1.28
azure-core==1.35.0
azure-identity==1.23.1
azure-iot-device==2.14.0
azure-iot-hub==2.6.1
azure-mgmt-core==1.6.0
azure-mgmt-iothub==4.0.0
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.5
deprecation==2.1.0
idna==3.10
isodate==0.7.2
janus==2.0.0
msal==1.33.0
msal-extensions==1.3.1
msrest==0.7.1
oauthlib==3.3.1
packaging==25.0
paho-mqtt==1.6.1
pycparser==2.22
PyJWT==2.10.1
PySocks==1.7.1
requests==2.32.4
requests-oauthlib==2.0.0
requests-unixsocket2==1.0.0
six==1.17.0
typing_extensions==4.14.1
uamqp==1.6.11
urllib3==2.5.0
